refactor: treat Svelte 4 as the exception
