+++ packages/adapter-sveltekit/public/AlternateGrid/typescript.svelte
@@ -2,7 +2,11 @@
 	import { PrismicImage, PrismicRichText } from "@prismicio/svelte";
 	import { type Content, isFilled } from "@prismicio/client";
 
-	export let slice: Content.PascalNameToReplaceSlice;
+	interface Props {
+		slice: Content.PascalNameToReplaceSlice;
+	}
+
+	let { slice }: Props = $props();
 </script>
 
 <section
