+++ packages/adapter-sveltekit/public/CustomerLogos/javascript.svelte
@@ -7,9 +7,9 @@
 	} from "@prismicio/svelte";
 
 	/**
-	 * @type {import("@prismicio/client").Content.${PascalNameToReplace}Slice}
+	 * @type {{ slice: import("@prismicio/client").Content.${PascalNameToReplace}Slice }}
 	 */
-	export let slice;
+	let { slice } = $props();
 </script>
 
 <section
