+++ packages/adapter-sveltekit/public/Hero/typescript.svelte
@@ -6,7 +6,11 @@
 	} from "@prismicio/svelte";
 	import { type Content, isFilled } from "@prismicio/client";
 
-	export let slice: Content.PascalNameToReplaceSlice;
+	interface Props {
+		slice: Content.PascalNameToReplaceSlice;
+	}
+
+	let { slice }: Props = $props();
 </script>
 
 <section
